+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/mobs/BeehemothEntity.java
@@ -31,6 +31,7 @@ import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.Difficulty;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.damagesource.DamageSource;
@@ -224,14 +225,16 @@ public class BeehemothEntity extends TamableAnimal implements FlyingAnimal, Sadd
                 if (entity != null && entity.getUUID().equals(getOwnerUUID())) {
                     addFriendship((int) (-3 * amount));
                 }
-                if (BzConfig.beehemothTriggersWrath && entity instanceof LivingEntity livingEntity) {
+
+                if (BzConfig.aggressiveBees &&
+                    BzConfig.beehemothTriggersWrath &&
+                    entity instanceof LivingEntity livingEntity)
+                {
                     addFriendship((int) (-amount));
 
-                    if (!(livingEntity instanceof Player player && player.isCreative()) &&
-                            (livingEntity.level.dimension().location().equals(Bumblezone.MOD_DIMENSION_ID) ||
-                            BzConfig.allowWrathOfTheHiveOutsideBumblezone) &&
-                            !livingEntity.isSpectator() &&
-                            BzConfig.aggressiveBees)
+                    if (!(livingEntity instanceof Player player && (player.isCreative() || level.getDifficulty() == Difficulty.PEACEFUL)) &&
+                        (livingEntity.level.dimension().location().equals(Bumblezone.MOD_DIMENSION_ID) || BzConfig.allowWrathOfTheHiveOutsideBumblezone) &&
+                        !livingEntity.isSpectator())
                     {
                         if(livingEntity.hasEffect(BzEffects.PROTECTION_OF_THE_HIVE)) {
                             livingEntity.removeEffect(BzEffects.PROTECTION_OF_THE_HIVE);