+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### **(V.6.6.3 Changes) (1.19.2 Minecraft)**
+
+##### Entities:
+Bee Queen will not get angry at attacking Players in Peaceful mode now.
+
+##### Effects:
+Wrath of the Hive will not apply to players in Peaceful mode now.
+
+
 ### **(V.6.6.2 Changes) (1.19.2 Minecraft)**
 
 ##### Entities: